Reorganize component structure for better maintainability
- Rename /components/filters/ to /filter-fields/ for clarity
- Rename filter field files to remove '-filters' suffix
- Rename data-explorer filter files to remove 'filter-' prefix
- Create /components/landing-page/ directory for homepage components
- Consolidate table components into /data-explorer/table/ subdirectory
- Move all chart components into /meta-analysis/charts/
- Remove unused Footer and LoadingDots components
- Delete unused /components/tables/ directory
- Update all imports to reflect new structure

This creates a more consistent feature-based organization where
components are grouped by the features they support (landing-page,
data-explorer, meta-analysis) rather than by component type.
