+++ components/data-explorer/table/cell-badges.tsx
