+++ components/data-explorer/table/cell-long-text.tsx
