--- components/footer.tsx
@@ -1,10 +1,0 @@
-export const Footer = () => {
-  return (
-    <footer
-      id="footer"
-      className="mx-auto flex grow items-end py-4 text-muted-foreground "
-    >
-      LIME 2025
-    </footer>
-  )
-}
