+++ components/landing-page/counter.tsx
