+++ components/landing-page/paper-carousel.tsx
