+++ components/meta-analysis/charts/highlight-bar-chart.tsx
